+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Added
 
  - Retries for Vault backends. `#12`_
  - Alternative constructors for ``Konfig``
+ - Add configuration extending. `#34`_
 
 0.6.0 - 2019-06-30
 ------------------
@@ -19,4 +20,5 @@ Added
 
 .. _Unreleased: https://github.com/kiwicom/konfetti/compare/0.6.0...HEAD
 
+.. _#34: https://github.com/kiwicom/konfetti/issues/34
 .. _#12: https://github.com/kiwicom/konfetti/issues/12